move tests to a more relevant location
